fixed compile errors in gps driver and renamed to Venus838FLPx
